requirements.txt
blinker==1.4
click==6.7
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1
psycopg2-binary==2.7.5
public==1.0.3
query-string==1.0.2
SQLAlchemy==1.2.8
Werkzeug==0.14.1
python-dateutil-2.8.2 six-1.16.0